Symbol display formatter
