+++ src/Raven.CodeAnalysis/Symbols/Symbol.cs
@@ -1,7 +1,6 @@
 using System.Collections.Immutable;
 using System.Diagnostics;
 using System.Reflection;
-using System.Text;
 
 using Raven.CodeAnalysis.Syntax;
 
@@ -102,141 +101,6 @@ internal abstract class Symbol : ISymbol
 
     public virtual bool IsStatic => false;
 
-    public string ToDisplayString(SymbolDisplayFormat format = default!)
-    {
-        format ??= SymbolDisplayFormat.CSharpErrorMessageFormat;
-
-        var result = new StringBuilder();
-
-        if (this is ILocalSymbol localSymbol)
-        {
-            if (format.LocalOptions.HasFlag(SymbolDisplayLocalOptions.IncludeType))
-            {
-                var localType = localSymbol.Type.ToDisplayString(format);
-                result.Append($"{localType} ");
-            }
-
-            result.Append(localSymbol.Name);
-            return result.ToString();
-        }
-
-        if (this is IParameterSymbol parameterSymbol)
-        {
-            if (format.ParameterOptions.HasFlag(SymbolDisplayParameterOptions.IncludeType))
-            {
-                var localType = parameterSymbol.Type.ToDisplayString(format);
-                result.Append($"{localType} ");
-            }
-
-            result.Append(parameterSymbol.Name);
-            return result.ToString();
-        }
-
-        // Example: Include namespace and type qualification
-        if (format.TypeQualificationStyle == SymbolDisplayTypeQualificationStyle.NameAndContainingTypesAndNamespaces)
-        {
-            // Assume `Namespace` and `ContainingType` are properties of the symbol
-            if (ContainingNamespace.ContainingNamespace is not null)
-            {
-                var ns = GetFullNamespace(this);
-                result.Append(ns).Append(".");
-            }
-
-            if (ContainingType is not null)
-            {
-                var type = GetFullType(this);
-                result.Append(type).Append(".");
-            }
-        }
-
-        // Append the symbol's name
-        result.Append(Name); // Assume `Name` is a property of the symbol
-
-        if (this is INamedTypeSymbol typeSymbol)
-        {
-            // Handle generics
-            if (format.GenericsOptions.HasFlag(SymbolDisplayGenericsOptions.IncludeTypeParameters) &&
-                typeSymbol.TypeParameters != null)
-            {
-                result.Append("<");
-                result.Append(string.Join(", ",
-                    typeSymbol.TypeParameters)); // Assume `TypeParameters` is a list of generic type names
-                result.Append(">");
-            }
-        }
-
-        if (this is IMethodSymbol methodSymbol)
-        {
-            // Return type (if requested)
-            if (format.MemberOptions.HasFlag(SymbolDisplayMemberOptions.IncludeType))
-            {
-                var returnType = methodSymbol.ReturnType?.ToDisplayString(format);
-                if (!string.IsNullOrEmpty(returnType))
-                {
-                    result.Insert(0, returnType + " ");
-                }
-            }
-
-            // Handle method parameters (if the symbol is a method)
-            if (format.DelegateStyle == SymbolDisplayDelegateStyle.NameAndSignature)
-            {
-                result.Append("(");
-                result.Append(string.Join(", ", methodSymbol.Parameters.Select(p => FormatParameter(p, format))));
-                result.Append(")");
-            }
-        }
-
-        // Example for accessibility modifiers
-        if (format.MemberOptions.HasFlag(SymbolDisplayMemberOptions.IncludeAccessibility))
-        {
-            if (this.DeclaredAccessibility is not Accessibility.NotApplicable)
-            {
-                result.Insert(0,
-                    this.DeclaredAccessibility.ToString().ToLower() + " "); // Assume `Accessibility` is a property
-            }
-        }
-
-        // Handle miscellaneous options
-        if (format.MiscellaneousOptions.HasFlag(SymbolDisplayMiscellaneousOptions.EscapeIdentifiers))
-        {
-            return EscapeIdentifier(result.ToString());
-        }
-
-        return result.ToString();
-    }
-
-    // Helper method to format a parameter
-    private string FormatParameter(IParameterSymbol parameter, SymbolDisplayFormat format)
-    {
-        var sb = new StringBuilder();
-
-        if (format.ParameterOptions.HasFlag(SymbolDisplayParameterOptions.IncludeModifiers))
-        {
-            //if (parameter.IsRef) sb.Append("ref ");
-            //if (parameter.IsOut) sb.Append("out ");
-        }
-
-        if (format.ParameterOptions.HasFlag(SymbolDisplayParameterOptions.IncludeType))
-        {
-            sb.Append(parameter.Type.ToDisplayString(format));
-            sb.Append(" ");
-        }
-
-        if (format.ParameterOptions.HasFlag(SymbolDisplayParameterOptions.IncludeName))
-        {
-            sb.Append(parameter.Name); // Assume `Name` is a property of the parameter
-        }
-
-        return sb.ToString();
-    }
-
-    // Helper method to escape identifiers
-    private string EscapeIdentifier(string identifier)
-    {
-        // Replace any special characters or keywords with escaped versions
-        return identifier; //identifier.Replace("<", "&lt;").Replace(">", "&gt;");
-    }
-
     public bool Equals(ISymbol? other, SymbolEqualityComparer comparer)
     {
         return comparer.Equals(this, other);
@@ -262,42 +126,10 @@ internal abstract class Symbol : ISymbol
         return Equals((ISymbol)other);
     }
 
-    private string GetFullNamespace(ISymbol symbol)
-    {
-        var namespaces = new List<string>();
-        var currentNamespace = symbol.ContainingNamespace;
-
-        // Traverse all containing namespaces
-        while (currentNamespace is not null && !currentNamespace.IsGlobalNamespace)
-        {
-            namespaces.Insert(0, currentNamespace.Name);
-            currentNamespace = currentNamespace.ContainingNamespace;
-        }
-
-        // Join namespaces with '.'
-        return string.Join(".", namespaces);
-    }
-
-    private string GetFullType(ISymbol symbol)
-    {
-        var types = new List<string>();
-        var currentType = symbol.ContainingType;
-
-        // Traverse all containing types
-        while (currentType is not null)
-        {
-            types.Insert(0, currentType.Name);
-            currentType = currentType.ContainingType;
-        }
-
-        // Join types with '.'
-        return string.Join(".", types);
-    }
-
-    private string GetDebuggerDisplay() => ToDisplayString(SymbolDisplayFormat.FullyQualifiedFormat);
+    private string GetDebuggerDisplay() => this.ToDisplayString(SymbolDisplayFormat.FullyQualifiedFormat);
 
     public override string ToString()
     {
-        return ToDisplayString();
+        return this.ToDisplayString();
     }
 }
